Add ServiceCallFactorySpec
