+++ product_code/shared/server/src/test/resources/application.conf
@@ -1,0 +1,2 @@
+# The key is "changeme" encoded in Base64, the Jwts.builder accepts this key encoding
+play.http.secret.key = "Y2hhbmdlbWU="
